Module Liquidsoap_core_formats.Content_base

type 'a chunk = {

data : 'a;

offset : int;

length : int;

}

type ('a, 'b) chunks = {

params : 'a;

mutable chunks : 'b chunk list;

mutable total_length : int;

}

module Contents : sig ... end

val merge_param : ?compare:??? -> name:string -> ('a option * 'a option) -> 'a option

val print_optional : (string * string option) list -> string

exception Invalid

exception Incompatible_format of Contents.format * Contents.format

module type FormatSpecs = sig ... end

What a content type is, as far as the language is concerned: its kind, its parameters and how they are named, parsed and printed.

module type DataSpecs = sig ... end

The buffers a content type holds, which only a process that streams needs.

module type ContentSpecs = sig ... end

module type Format = sig ... end

module type Content = sig ... end

type data = Contents.data

type kind = Contents.kind

type kind_handler = {

default_format : unit -> Contents.format;

string_of_kind : unit -> string;

}

val max_contents : int

val dummy_kind_handler_fn : Contents.kind_content -> kind_handler

val kind_handler_fns : (Contents.kind_content -> kind_handler) array

val get_kind_handler : Contents.kind_content Contents.entry -> kind_handler

val kind_parsers : (string -> Contents.kind option) Stdlib.Queue.t

exception Parsed_kind of kind

val kind_of_string : string -> kind

type format = Contents.format

type format_handler = {

kind : unit -> kind;

string_of_format : unit -> string;

merge : format -> unit;

compatible : format -> bool;

duplicate : unit -> format;

serialize : unit -> string;

}

val dummy_format_handler_fn : Contents.format_content Unifier.t -> format_handler

val format_handler_fns : (Contents.format_content Unifier.t -> format_handler) array

val get_format_handler : Contents.format_content Unifier.t Contents.entry -> format_handler

val format_param_parsers : (string -> string -> Contents.format option) array

val format_parsers : (string -> Contents.format option) array

val format_make_fns : (Contents.format -> int option -> Contents.data) array

val parse_param : 'a Contents.entry -> string -> string -> Contents.format

type data_handler = {

sub : data -> int -> int -> data;

truncate : data -> int -> data;

_length : data -> int;

is_empty : data -> bool;

copy : data -> data;

checksum : data -> string;

format : data -> format;

append : data -> data -> data;

}

val dummy_handler : data_handler

val data_handlers : data_handler array

val register_data_handler : int -> data_handler -> unit

val get_data_handler : (int * 'a) -> data_handler

val make : ?length:??? -> Contents.format -> Contents.data

val sub : data -> int -> int -> data

val truncate : data -> int -> data

val is_empty : data -> bool

val length : data -> int

val append : data -> data -> data

val copy : data -> data

val checksum : data -> string

val format : data -> format

val kind : Contents.format_content Unifier.t Contents.entry -> kind

val default_format : Contents.kind_content Contents.entry -> Contents.format

val string_of_format : Contents.format_content Unifier.t Contents.entry -> string

val merge : Contents.format -> format -> unit

val duplicate : Contents.format_content Unifier.t Contents.entry -> format

val compatible : Contents.format_content Unifier.t Contents.entry -> format -> bool

val string_of_kind : Contents.kind_content Contents.entry -> string

val serialize_format : Contents.format_content Unifier.t Contents.entry -> string

val parse_format : string -> Contents.format option

val content_names : string list Stdlib.ref

type content_lang_spec = {

format_name : string;

method_name : string;

content_typ : Liquidsoap_lang_types.Type.t;

format_to_value : Contents.format -> Liquidsoap_lang_values.Value.t;

}

val content_lang_specs : content_lang_spec Stdlib.Queue.t

val on_content_lang : (content_lang_spec -> unit) Stdlib.ref

val register_content_lang : string -> string -> Liquidsoap_lang_types.Type.t -> (Contents.format -> Liquidsoap_lang_values.Value.t) -> unit

module MkFormatBase (F : FormatSpecs) : Format with type kind = F.kind and type params = F.params

module MkDataBase (F : Format) (D : DataSpecs with type params = F.params) : Content with type kind = F.kind and type params = F.params and type data = D.data

module MkContentBase (C : ContentSpecs) : Content with type kind = C.kind and type params = C.params and type data = C.data