Module Liquidsoap_core_formats.Frame_type

val default_audio_channels : (unit -> int) Stdlib.ref

What a frame gets when a script gives it no type of its own, which core reads from the settings.

val default_video : (unit -> bool) Stdlib.ref

val make : ?pos:Liquidsoap_lang_prelude.Pos.t -> Liquidsoap_lang_types.Type.t -> Liquidsoap_lang_types.Type.t Fields.t -> Liquidsoap_lang_types.Type.t

val internal_tracks : ?pos:Liquidsoap_lang_prelude.Pos.t -> unit -> Liquidsoap_lang_types.Type.t

val pcm_audio : ?pos:Liquidsoap_lang_prelude.Pos.t -> unit -> Liquidsoap_lang_types.Type.t

val set_field : Liquidsoap_lang_types.Type.t -> Fields.field -> Liquidsoap_lang_types.Type.t -> Liquidsoap_lang_types.Type.t

val get_field : Liquidsoap_lang_types.Type.t -> Fields.field -> Liquidsoap_lang_types.Type.t

val get_fields : Liquidsoap_lang_types.Type.t -> Fields.field list

val content_type : Liquidsoap_lang_types.Type.t -> Content_base.format Fields.t