Module Liquidsoap_ffmpeg_formats.Ffmpeg_encoder_type

type codecs = {

media_type : string -> [ `Audio | `Video | `Subtitle ] option;

channels_of_layout : string -> int option;

}

What only libav can answer about a track. A process without it reads the conventions in the track's name and parameters, and leaves the rest unknown.

val codecs : codecs option Stdlib.ref

val implement : codecs -> unit

val named_layouts : (string * int) list

val channels_of_layout : string -> int option

val contains : string -> string -> bool

val term_string : [> `String of 'a ] Liquidsoap_lang_ast.Term.term -> 'a option

val layout_name : [> `Float of float | `String of string ] Liquidsoap_lang_ast.Term.term -> string option

val has_content : static_string:('a -> 'b option) -> 'b -> (string * 'a) list -> bool

val media_type : static_string:('a -> string option) -> string -> (string * 'a) list -> [ `Audio | `Subtitle | `Video ] option

val static_channels : string -> Liquidsoap_lang_ast.Term.ast Liquidsoap_lang_ast.Term.term -> int

val channels : (string * Liquidsoap_lang_ast.Term.ast Liquidsoap_lang_ast.Term.term) list -> int

val pcm_kind : static_string:('a -> string option) -> (string * 'a) list -> Liquidsoap_core_formats__Content_base.Contents.kind_content Liquidsoap_core_formats__Content_base.Contents.entry

val parse_encoder_name : string -> string * [> `Copy | `Drop | `Internal | `Raw ]

val unknown_track : unit -> Liquidsoap_lang_types.Type.t

val kind : string -> Liquidsoap_core_formats.Content_base.Contents.format

val format_of_track : [< `Copy | `Drop | `Internal | `Raw ] -> string -> (string * Liquidsoap_lang_ast.Term.ast Liquidsoap_lang_ast.Term.term) list -> Liquidsoap_lang_types.Type.t

val type_of_encoder : [> `Encoder of string * [< `Anonymous of string | `Encoder of 'a | `Labelled of string * Liquidsoap_lang_ast.Term.t ] list ] list -> (int, Liquidsoap_lang_types.Type.t) Liquidsoap_lang_data__Methods.t