Module Dtools

ocaml-dtools.

author Stephane Gimenez

module Conf : sig ... end

Configuration management module.

module Init : sig ... end

Initialisation management module. Allow to define procedures that must be executed at start up, and procedures that are to be executed at exit to have a clean quit.

module Log : sig ... end