Module Dtools.Log

type entry = {

time : float;

label : string option;

level : int option;

log : string;

}

type t = < active : int -> bool ; level : int ; set_level : int -> unit ; path : Conf.path ; f : 'a. int -> ('a, unit, string, unit) Stdlib.format4 -> 'a ; g : 'a. ?colorize:(entry -> entry) -> int -> ('a, unit, string, unit) Stdlib.format4 -> 'a >

Type for loggers.

type custom_log = {

timestamp : bool;

exec : string -> unit;

}

val add_custom_log : string -> custom_log -> unit

Add a custom logging functions.

val rm_custom_log : string -> unit

Remove a custom logging functions.

val make : Conf.path -> t

Make a logger labeled according to the given path.

val start : Init.t

An atom that starts the logging.

val stop : Init.t

An atom that stops the logging.

val conf : Conf.ut

val conf_level : int Conf.t

val conf_unix_timestamps : bool Conf.t

val conf_stdout : bool Conf.t

val conf_file : bool Conf.t

val conf_file_path : string Conf.t

val conf_file_append : bool Conf.t

val conf_file_perms : int Conf.t

val args : (string list * Stdlib.Arg.spec * string) list

A set of command line options to be used with the Arg module.