Module Monad.Mutex

This module implements monadic * mutex computations. They can be used * to write blocking code that is compatible * with duppy's tasks, i.e. Mutex.lock m blocks * the calling computation and not the calling thread.

module type Mutex_control = sig ... end

Information used to initialize a Mutex module. * priority and scheduler are used to initialize a task * which treat mutexes as well as conditions from the below * Condition module.

module type Mutex_t = sig ... end

module Factory (_ : Mutex_control) : Mutex_t