Module type Mutex.Mutex_t

type mutex

Type for a mutex.

module Control : Mutex_control

val create : unit -> mutex

create () creates a mutex.

val lock : mutex -> (unit, 'a) t

A computation that locks a mutex * and returns unit afterwards. Computation * will be blocked until the mutex is successfully locked.

val try_lock : mutex -> (bool, 'a) t

A computation that tries to lock a mutex. * Returns immediately true if the mutex was successfully locked * or false otherwise.

val unlock : mutex -> (unit, 'a) t

A computation that unlocks a mutex. * Should return immediately.