Module AFrame

Specialization of Frame for content kinds restricted to audio, * with a fixed number of channels. * All units are in audio samples instead of main clock ticks.

Types

type t = Frame.t

Basic manipulation

val duration : unit -> float

Duration in seconds.

Track marks

* Track marks are track limits. * Everything below is in samples.

val size : unit -> int

Size of an audio frame.

val position : t -> int

Current position in frame.

Helpers

val pcm : t -> Content.Audio.data

Get PCM content. Raises Content.Invalid * if frame content is not pcm.

val s16le : t -> string

Get content in S16LE format. Raises Content.Invalid * if frame content is not pcm.

Sound processing

val rms : t -> int -> int -> float array

RMS (root mean square) of a portion of a frame.