Module Charset

type t

A character set.

exception Unknown_encoding of string

exception Unsupported_encoding of t

val ascii : t

val latin1 : t

val utf8 : t

val utf16 : t

val utf16be : t

val utf16le : t

val utf32 : t

val utf32be : t

val utf32le : t

val ucs4 : t

val of_string : string -> t

Charset from string.

val to_string : t -> string

String name of charset.

val convert : ?fail:bool -> ?source:t -> ?target:t -> string -> string

Convert between charsets. By default, source charset is automatically detected and target charset is UTF8. The function will silently fail unless fail is set to true.