Module Decoder

val log : Log.t

type file = string

type stream = string

type input = {

read : bytes -> int -> int -> int;

lseek : (int -> int) option;

tell : (unit -> int) option;

length : (unit -> int) option;

}

type fps = {

num : int;

den : int;

}

type buffer = {

generator : Generator.t;

put_pcm : ?field:Frame.field -> samplerate:int -> Mm.Audio.t -> unit;

put_yuva420p : ?field:Frame.field -> fps:fps -> Mm.Video.Canvas.image -> unit;

}

type decoder = {

decode : buffer -> unit;

eof : buffer -> unit;

seek : int -> int;

close : unit -> unit;

}

type file_decoder_ops = {

fread : int -> Frame.t;

remaining : unit -> int;

fseek : int -> int;

fclose : unit -> unit;

}

type stream_decoder = input -> decoder

type image_decoder = {

image_decoder_priority : unit -> int;

check_image : file -> bool;

decode_image : file -> Mm.Video.Image.t;

}

type file_decoder = metadata:Frame.metadata -> ctype:Frame.content_type -> string -> file_decoder_ops

type decoder_specs = {

priority : unit -> int;

file_extensions : unit -> string list option;

mime_types : unit -> string list option;

file_type : metadata:Frame.metadata -> ctype:Frame.content_type -> string -> Frame.content_type option;

file_decoder : file_decoder option;

stream_decoder : (ctype:Frame.content_type -> string -> stream_decoder) option;

}

val decoders : decoder_specs Plug.t

val conf_decoder : Dtools.Conf.ut

val conf_mime_types : Dtools.Conf.ut

val conf_file_extensions : Dtools.Conf.ut

val conf_priorities : Dtools.Conf.ut

val conf_image_priorities : Dtools.Conf.ut

val openfile : string -> Unix.file_descr

Open file with readonly, cloexec and share delete on windows.

val test_file : log:Log.t -> extension:string option -> mime:stream -> mimes:stream list option -> extensions:string list option -> stream -> bool

Test file extension and mime if available

val can_decode_type : Frame.content_type -> Frame.content_type -> bool

Test if we can decode for a content_type. This include cases where we know how to convert channel layout.

val get_file_decoder : metadata:Frame.metadata -> ctype:Frame.content_type -> string -> (string * (unit -> file_decoder_ops)) option

val get_stream_decoder : ctype:Frame.content_type -> string -> stream_decoder option

val image_file_decoders : image_decoder Plug.t

val check_image_file_decoder : file -> bool

val get_image_file_decoder : file -> Mm.Video.Image.t

val mk_buffer : ctype:Frame.content_type -> Generator.t -> buffer

val file_decoder : filename:string -> remaining:(unit -> int) -> ctype:Frame.content_type -> decoder -> file_decoder_ops

val opaque_file_decoder : filename:string -> ctype:Frame.content_type -> (input -> decoder) -> file_decoder_ops