Module External_decoder

Decode files using an external decoder.

First, an external decoder that receives * on its stdin.

val on_stderr : (bytes -> int -> int -> int) -> [> `Continue ]

val external_input : string -> Decoder.input -> Decoder.input * (unit -> unit)

This function is used to wrap around the "real" input. * It pipes its data to the external process and read * the available output.

val duration : string -> float

val create : string -> Frame.content_type -> string -> Decoder.file_decoder_ops

A function to wrap around the Wav_aiff_decoder

val create_stream : string -> Decoder.input -> Decoder.decoder

val audio_n : int -> Content_base.Contents.format

val test_ctype : ('a -> int) -> 'a -> Content_base.Contents.format Frame.Fields.t option

val register_stdin : name:string -> doc:string -> priority:int -> mimes:string list option -> file_extensions:string list option -> test:(string -> int) -> string -> unit

Now an external decoder that directly operates * on the file. The remaining time in this case * can only be approximative. It is -1 while * the file is being decoded and the length * of the buffer when the external decoder * has exited.

val log : Log.t

val external_input_oblivious : ('a -> string) -> 'a -> float -> Decoder.file_decoder_ops

val register_oblivious : name:string -> doc:string -> priority:int -> mimes:string list option -> file_extensions:string list option -> test:(string -> int) -> process:(string -> string) -> float -> unit