Module Frame

Operations on frames, which are small portions of streams.

Frame definitions

module Fields : sig ... end

type field = Fields.field

type content_type = Content.format Fields.t

Precise description of the channel types for the current track.

module Metadata : sig ... end

type metadata = Metadata.t

Metadata of a frame.

type t = Content.data Fields.t

A frame is a chunk of data which should be at most size

Content-independent frame operations

All units are in ticks (main clock).

val create : length:int -> content_type -> t

Create a frame of a given content type.

val slice : t -> int -> t

Get at most length data from the start of the given frame.

val after : t -> int -> t

Get the content after a given offset.

val sub : t -> int -> int -> t

sub frame ofs len: get the a subset of length len of the frame content, starting at ofs.

val content_type : t -> content_type

Get a frame's content type.

val append : t -> t -> t

append f f' appends the data from f' to f. f' can possibly have more fields that f.

val get : t -> field -> Content.data

Get a frame's content.

val set : t -> field -> Content.data -> t

Set a frame's content.

val set_data : t -> field -> (?offset:int -> ?length:int -> 'a -> Content.data) -> 'a -> t

Set a frame's content using data. Data is assumed to be of the frame's position length. The type is designed to work with `Content.*.lift_data` functions.

val audio : t -> Content.data

Get a frame's audio content.

val video : t -> Content.data

Get a frame's video content.

val midi : t -> Content.data

Get a frame's midi content.

val position : t -> int

val remaining : t -> int

Remaining data length in the frame.

val is_partial : t -> bool

Is the frame partially filled, i.e. is its end position strictly before Lazy.force Frame.size?

Track marks

val track_marks : t -> int list

List of track marks in a frame.

val add_track_mark : t -> int -> t

Add a track mark to a frame

val add_track_marks : t -> int list -> t

Add a multiple track marks to a frame

val has_track_marks : t -> bool

true is frame has a track mark.

val has_track_mark : t -> int -> bool

true is frame has a track mark at the given position

val drop_track_marks : t -> t

Remove all track marks from the frame.

Metadata

exception No_metadata

val add_metadata : t -> int -> metadata -> t

Attach metadata at a given position to the frame.

val free_metadata : t -> int -> t

val get_metadata : t -> int -> metadata option

Retrieve metadata at a given position.

val get_all_metadata : t -> (int * metadata) list

Retrieve all metadata.

val add_all_metadata : t -> (int * metadata) list -> t

Attach multiple metadata to a frame.

val map_metadata : t -> ((int * metadata) -> (int * metadata) option) -> t

Map a function over the frame's metadata.

Content operations

val string_of_content_type : content_type -> string

val assert_compatible : content_type -> content_type -> unit

val compatible : content_type -> content_type -> bool

Format settings

The channel numbers are only defaults, used when channel numbers * cannot be inferred / are not forced from the context. * I'm currently unsure how much they are really useful.

val audio_channels : int Stdlib.Lazy.t

Default number of audio channels.

val default_video_enabled : bool Stdlib.Lazy.t

Is video enabled?

val midi_channels : int Stdlib.Lazy.t

Default number of MIDI channels.

val video_width : int Stdlib.Lazy.t

Width of video images.

val video_height : int Stdlib.Lazy.t

Height of video images.

val audio_rate : int Stdlib.Lazy.t

Rate of audio (in samples per second).

val video_rate : int Stdlib.Lazy.t

Video rate (in images per second).

val midi_rate : int Stdlib.Lazy.t

val main_rate : int Stdlib.Lazy.t

Ticks per second.

val size : int Stdlib.Lazy.t

val duration : float Stdlib.Lazy.t

Duration of a frame in seconds.

Time and frequency conversions

Conversion between the internal unit (main ticks), seconds, and data units.

val audio_of_main : int -> int

Duration of given number of samples in ticks.

val video_of_main : int -> int

val video_of_main_f : int -> float

val midi_of_main : int -> int

val main_of_audio : int -> int

val main_of_video : int -> int

val main_of_midi : int -> int

val main_of_seconds : float -> int

val audio_of_seconds : float -> int

val video_of_seconds : float -> int

val seconds_of_main : int -> float

val seconds_of_audio : int -> float

val seconds_of_video : int -> float