Module Frame_base

Operations on frames, which are small portions of streams.

Frame definitions

val field_idx : int Stdlib.Atomic.t

module FieldNames : sig ... end

module Fields : sig ... end

type field = Fields.field

type content_type = Content_base.format Fields.t

Precise description of the channel types for the current track.

type t = Content_base.data Fields.t

val position : ('a, Content_base.data) Liquidsoap_lang__Methods.t -> int

val remaining : ('a, Content_base.data) Liquidsoap_lang__Methods.t -> int

val is_partial : ('a, Content_base.data) Liquidsoap_lang__Methods.t -> bool

module Metadata = Metadata_base

Metadata of a frame.

type metadata = Metadata_base.t

val audio_format : pcm_kind:Content_base.Contents.kind -> Content_audio.params -> Content_base.Contents.format

val format_of_channels : pcm_kind:Content_base.Contents.kind -> int -> Content_base.Contents.format

val add_timed_content : ?length:int -> (int, Content_base.data) Liquidsoap_lang__Methods.t -> (int, Content_base.data) Liquidsoap_lang__Methods.t