Module Frame_settings

Configuration entries

val log : Log.t

module Conf = Dtools.Conf

val conf : Conf.ut

val conf_duration : float Conf.t

val conf_audio : Conf.ut

val conf_audio_samplerate : int Conf.t

val conf_audio_channels : int Conf.t

val conf_audio_size : int Conf.t

val conf_video : Conf.ut

val conf_video_default : bool Conf.t

val conf_video_framerate : int Conf.t

val conf_video_width : int Conf.t

val conf_video_height : int Conf.t

val conf_midi : Conf.ut

val conf_midi_channels : int Conf.t

Format parameters

val lazy_config_eval : bool Stdlib.ref

val delayed_eval : (unit -> unit) Stdlib.Queue.t

val delayed : (unit -> 'a) -> 'a Stdlib.Lazy.t

val delayed_conf : to_string:('a -> string) -> < get : 'a ; ut : Conf.ut.. > -> 'a Stdlib.Lazy.t

val (!!) : 'a Stdlib.Lazy.t -> 'a

The channel numbers are only defaults, used when channel numbers * cannot be inferred / are not forced from the context. * I'm currently unsure how much they are really useful.

val audio_channels : int Stdlib.Lazy.t

val default_video_enabled : bool Stdlib.Lazy.t

val midi_channels : int Stdlib.Lazy.t

val video_width : int Stdlib.Lazy.t

val video_height : int Stdlib.Lazy.t

val audio_rate : int Stdlib.Lazy.t

val video_rate : int Stdlib.Lazy.t

val midi_rate : int Stdlib.Lazy.t

val gcd : int -> int -> int

Greatest common divisor.

val lcm : int -> int -> int

Least common multiplier.

val upper_multiple : int -> int -> int

upper_multiple k m is the least multiple of k that is >=m.

val main_rate : int Stdlib.Lazy.t

The main clock is the slowest possible that can convert to both * the audio and video clocks.

val m_o_a : int Stdlib.Lazy.t

Precompute those ratios to avoid too large integers below.

val m_o_v : int Stdlib.Lazy.t

val main_of_audio : int -> int

val main_of_video : int -> int

val main_of_midi : int -> int

val audio_of_main : int -> int

val video_of_main : int -> int

val video_of_main_f : int -> float

val midi_of_main : int -> int

val main_of_seconds : float -> int

val audio_of_seconds : float -> int

val video_of_seconds : float -> int

val seconds_of_main : int -> float

val seconds_of_audio : int -> float

val seconds_of_video : int -> float

val size : int Stdlib.Lazy.t

The frame size (in main ticks) should allow for an integer * number of samples of all types (audio, video). * With audio@44100Hz and video@25Hz, ticks=samples and one video * sample takes 1764 ticks: we need frames of size N*1764.

val duration : float Stdlib.Lazy.t