Module Generator

A generator is a collection of content with different length. The set of content contained by a generator is given by its content_type to which metadata and track marks are always added.

For now, metadata and track breaks content have infinite length. All other content have finite length.

A generator buffered length is the largest content length, excluding metadata and track breaks. If that length exceeds max_length, the generator is truncated to keep it under that value.

A generator length is the smallest content length, again excluding metadata and track breaks. This is the maximum content length that can be taken out of the generator.

type t

val create : ?log:Log.t -> ?max_length:int -> ?length:int -> ?content:Frame.t -> Frame_base.content_type -> t

val max_length : t -> int option

val set_max_length : t -> int option -> unit

val content_type : t -> Frame_base.content_type

val get_field : t -> Frame_base.field -> Content.data

val set_field : t -> Frame_base.field -> Content.data -> unit

val field_length : t -> Frame_base.field -> int

val length : t -> int

val buffered_length : t -> int

val remaining : t -> int

val truncate : t -> int -> unit

val keep : t -> int -> unit

val slice : t -> int -> Frame.t

val clear : t -> unit

val put : t -> Frame_base.field -> Content.data -> unit

val peek : t -> Content.data Frame_base.Fields.t

val peek_media : t -> Content.data Frame_base.Fields.t

val add_metadata : ?pos:int -> t -> Frame_base.metadata -> unit

val add_track_mark : ?pos:int -> t -> unit

val append : ?offset:int -> ?length:int -> t -> Frame_base.t -> unit