Class Icecast2.output

Sending encoded data to a shout-compatible server. * It directly takes the Lang param list and extracts stuff from it.

inherit Output.encoded

In this operator, we don't exactly follow the start/stop * mechanism of Output.encoded because we want to control * in a more subtle way the connection/disconnection with * icecast. * So we have specific icecast_start/stop procedures that * only deal with the shout connection. * And the global output_start/stop also deal with the encoder. * As a result, if shout gets disconnected, encoding will keep * going, and the sending will keep being attempted, which * will at some point trigger a restart.

val mutable restart_time : float

Time after which we should attempt to connect.

val mutable dump : Stdlib.out_channel option

File descriptor where to dump.

val mutable encoder : Encoder.encoder option

method encode : Frame.t -> Strings.t

method insert_metadata : Frame.Metadata.Export.t -> unit

method icecast_send : Strings.t -> unit

method send : Strings.t -> unit

method start : unit

method stop : unit

method icecast_start : unit

method icecast_stop : unit