Module Lifecycle

Module to expose and hook up actions that need to happen over the lifecycle of the application.

val on_load : name:string -> (unit -> unit) -> unit

Initial load. Used for registering stuff that needs all OCaml modules to have been registered.

val load : unit -> unit

Initialization

Initialization proceeds in 3 phases:

Init

Script parse

Application start

Init start

val init : unit -> unit

val before_init : name:string -> (unit -> unit) -> unit

val on_init : name:string -> (unit -> unit) -> unit

val after_init : name:string -> (unit -> unit) -> unit

Script parse

val before_script_parse : name:string -> (unit -> unit) -> unit

val on_script_parse : name:string -> (unit -> unit) -> unit

val after_script_parse : name:string -> (unit -> unit) -> unit

Application start

val before_start : name:string -> (unit -> unit) -> unit

val on_start : name:string -> (unit -> unit) -> unit

val after_start : name:string -> (unit -> unit) -> unit

val before_main_loop : name:string -> (unit -> unit) -> unit

Application main loop

val on_main_loop : name:string -> (unit -> unit) -> unit

val after_main_loop : name:string -> (unit -> unit) -> unit

Shutdown

Shutdown proceeds in 3 phases:

Core shutdown

Scheduler shutdown

Final cleanup

Stop

Core shutdown

val before_core_shutdown : name:string -> (unit -> unit) -> unit

val on_core_shutdown : name:string -> (unit -> unit) -> unit

val after_core_shutdown : name:string -> (unit -> unit) -> unit

Scheduler shutdown

val before_scheduler_shutdown : name:string -> (unit -> unit) -> unit

val on_scheduler_shutdown : name:string -> (unit -> unit) -> unit

val after_scheduler_shutdown : name:string -> (unit -> unit) -> unit

Final cleanup

val before_final_cleanup : name:string -> (unit -> unit) -> unit

val on_final_cleanup : name:string -> (unit -> unit) -> unit

val after_final_cleanup : name:string -> (unit -> unit) -> unit