Module Max_duration

class max_duration : override_meta:string -> duration:int -> Source.source -> object ... end

This one is a bit tricky as we want to make sure * that the underlying source is cleaned up when it's done * pulling. Used in switch-based transitions to avoid infinite * stack of sources.