Class Max_duration.max_duration

This one is a bit tricky as we want to make sure * that the underlying source is cleaned up when it's done * pulling. Used in switch-based transitions to avoid infinite * stack of sources.

inherit Source.operator

val mutable remaining : int

val mutable s : Source.source

method self_sync : Clock.self_sync

method fallible : bool

method private can_generate_frame : bool

method abort_track : unit

method remaining : int

method seek : int -> int

method seek_source : Source.source

method private check_for_override : Frame.t -> unit

method private generate_frame : Frame.t