Class Noblank.base

val state : [ `Blank of int | `Noise of int ] Stdlib.Atomic.t

State can be either

`Noise l: the source is considered to be emitting, but it has been silent for l samples;

`Blank l: the source is considered to be silent, but it has been noisy for l samples.

val dB_levels : float array option Stdlib.Atomic.t

method dB_levels : float array option

method private virtual log : Log.t

method is_blank : bool

method private string_of_state : [ `Blank of int | `Noise of int ] -> string

method private set_state : [ `Blank of int | `Noise of int ] -> unit

method private check_blank : Frame.t -> unit

This method should be called after the frame s has been filled, where p0 is the position in s before filling.