Module Pipe_output

base class

val output : Lang.module_name

val encoder_factory : ?format:Encoder.format -> Lang.value -> string -> Frame.Metadata.Export.t -> Encoder.encoder

val base_proto : (string * Lang.t * Lang.value option * string option) list

class virtual base : ?clock:Clock.t option -> source:Lang.value -> name:string -> (string * Lang.value) list -> object ... end

val url_proto : Lang.t -> (string * Lang.t * Lang.value option * string option) list

url output: discard encoded data, try to restart on encoding error (can be networking issues etc.)

class url_output : (string * Lang.value) list -> object ... end

Piped virtual class: open/close pipe, implements metadata interpolation and takes care of the various reload mechanisms.

val default_reopen_on_error : Lang.value

val default_reopen_on_metadata : Lang.value

val default_reopen_when : Lang.value

val pipe_meth : (string * Lang.scheme * string * (< abort_track : unit ; active : bool ; add_watcher : Source.watcher -> unit ; buffer : Generator.t ; cleanup_pipe : unit ; clock : Clock.t ; close_encoder : Strings.t ; close_pipe : unit ; consumed : int -> unit ; content_type : Frame.content_type ; content_type_computation_allowed : unit ; create_encoder : unit ; duration : int ; elapsed : int ; empty_frame : Frame.t ; encode : Frame.t -> Strings.t ; end_of_track : Frame.t ; fallible : bool ; force_sleep : unit ; frame_audio_position : int ; frame_has_track_mark : bool ; frame_metadata : (int * Frame.Metadata.t) list ; frame_position : int ; frame_track_mark : int option ; frame_type : Type.t ; get_frame : Frame.t ; get_mutable_content : Frame.field -> Content.data ; get_mutable_frame : Frame.field -> Frame.t ; get_partial_frame : (Frame.t -> Frame.t) -> Frame.t ; has_content_type : bool ; id : string ; insert_metadata : Frame.Metadata.Export.t -> unit ; interpolate : ?subst:(string -> string) -> string -> string ; is_open : bool ; is_ready : bool ; is_up : bool ; last_image : Frame.Fields.field -> Mm.Video.Canvas.image ; last_metadata : Frame.metadata option ; log : < active : int -> bool ; critical : 'a. ('a, unit, string, unit) Stdlib.format4 -> 'a ; debug : 'a. ('a, unit, string, unit) Stdlib.format4 -> 'a ; f : 'a. int -> ('a, unit, string, unit) Stdlib.format4 -> 'a ; important : 'a. ('a, unit, string, unit) Stdlib.format4 -> 'a ; info : 'a. ('a, unit, string, unit) Stdlib.format4 -> 'a ; level : int option ; set_level : int -> unit ; severe : 'a. ('a, unit, string, unit) Stdlib.format4 -> 'a > ; need_reopen : unit ; on_after_streaming_cycle : (unit -> unit) -> unit ; on_before_streaming_cycle : (unit -> unit) -> unit ; on_metadata : (Frame.metadata -> unit) -> unit ; on_sleep : (unit -> unit) -> unit ; on_track : (Frame.metadata -> unit) -> unit ; on_wake_up : (unit -> unit) -> unit ; open_pipe : unit ; output : unit ; peek_frame : Frame.t ; pos : Pos.Option.t ; prepare_pipe : unit ; register_command : ?usage:string -> descr:string -> string -> (string -> string) -> unit ; remaining : int ; reopen : unit ; reset : unit ; reset_last_metadata_on_track : bool ; seek : int -> int ; seek_source : Source.source ; self_sync : [ `Dynamic | `Static ] * Clock.sync_source option ; send : Strings.t -> unit ; set_frame_data : 'a. Frame.field -> (?offset:int -> ?length:int -> 'a -> Content.data) -> 'a -> Frame.t ; set_id : ?definitive:bool -> string -> unit ; set_name : string -> unit ; set_reset_last_metadata_on_track : bool -> unit ; set_stack : Pos.t list -> unit ; sleep : unit ; source_sync : bool -> Clock.sync_source option ; source_type : Source.source_type ; stack : Pos.t list ; stack_unifier : Pos.t list Unifier.t ; start : unit ; state : Start_stop.state ; stop : unit ; transition_to : Start_stop.state -> unit ; wake_up : unit ; write_pipe : string -> int -> int -> unit > -> Lang.value)) list

class virtual piped_output : ?clock:Clock.t option -> name:string -> (string * Lang.value) list -> object ... end

val chan_proto : Lang.t -> string -> (string * Lang.t * Lang.value option * string option) list

Out channel virtual class: takes care of current out channel and writing to it.

class virtual 'a chan_output : (string * Lang.value) list -> object ... end

File output

class virtual 'a file_output_base : (string * Lang.value) list -> object ... end

class file_output : ?clock:Clock.t option -> format_val:Lang.value -> (string * Lang.value) list -> object ... end

class file_output_using_encoder : ?clock:Clock.t option -> format_val:Lang.value -> (string * Lang.value) list -> object ... end

val file_proto : Lang.t -> (string * Lang.t * Lang.value option * string option) list

val new_file_output : ?clock:Clock.t -> (string * Lang.value) list -> piped_output

val output_file : Lang.module_name

External output

class external_output : ?clock:Clock.t option -> (string * Lang.value) list -> object ... end

val pipe_proto : Lang.t -> string -> (string * Lang.t * Lang.value option * string option) list