Module Playlist_parser

Plug for playlist parsing, in which src/playlists plugins come.

module Http = Liq_http

val log : Log.t

val conf_playlist : Dtools.Conf.ut

val conf_cue_in_metadata : string Dtools.Conf.t

val conf_cue_out_metadata : string Dtools.Conf.t

type playlist = ((string * string) list * string) list

A playlist is list of metadatas,uri

type parser = ?pwd:string -> string -> playlist

type plugin = {

strict : bool;

parser : parser;

}

A plugin is a boolean and a parsing function

val parsers : plugin Plug.t

Parsers are given a string and return a list of metadatas,uri, if possible.

val get_file : ?pwd:string -> string -> string

exception Exit of string * playlist

val search_valid : ?pwd:string -> string -> string * playlist

Get a valid parser for string. * The validity is not based on file type but only on success of the * parser instantiation. * Being based on file extension is weak, and troublesome when accessing a * remote file -- that would force us to create a local temporary file with the * same extension.