Module Process_handler

Asynchronous process handler. This module wraps around Unix.open_process_full * and Unix.select to manipulate external processes. Its API can be used in two * different ways: * - Provide callbacks for reading and writing to the process' standard outputs. * These callbacks are executed when Unix.select returns their respective * file descriptors as available for writing/reading. Thus, read and write * inside these callbacks should not be blocking although Unix.select * documentation stipulates that "large" writes may still be blocking. * - Use direct, blocking, read/write on the process' standard output. * in this case, reading and writing may be blocking, following the * usual Unix.write and Unix.read semantics. * The API abstracts the file descriptors and provides writing/reading * callbacks instead. Please note that there is no way to detect when the * process has exited unless reaching a zero-length read in its standard * output or error. Also, the default on_stdout and on_stderr callbacks * do not read from these. Therefore it is important to read all data coming * from at least one of these two outputs in order to make sure that the module * properly detects when the process has exited.

type t

A process.

type continuation = [

| `Continue

| `Stop
(*
triggers an asynchronous process stop, closing the process' stdin and waiting for termination
*)

| `Kill
(*
kill the process immediately
*)

| `Delay of float
(*
wait for a given amount of seconds
*)

| `Reschedule of Tutils.priority
(*
Update the process' priority and continue processing.
*)

]

Type for continuation decision.

type 'a callback = 'a -> continuation

A call back.

type pull = Stdlib.Bytes.t -> int -> int -> int

Function for reading data.

type push = Stdlib.Bytes.t -> int -> int -> int

Function for writing data.

type status = [

| `Exception of exn

| `Status of Unix.process_status

]

exception Finished

Trying to performed an operation on a stopped process.

val run : ?priority:Tutils.priority -> ?env:string array -> ?on_start:push callback -> ?on_stdin:push callback -> ?on_stdout:pull callback -> ?on_stderr:pull callback -> ?on_stop:(status -> bool) -> ?log:(string -> unit) -> string -> t

Create a process handler. Decisions returned by the callbacks are applied synchronously, i.e. when returning `Stop, the process' stdin is closed immediately after the callback has returned. The process is restarted when the function on_stop returns true.

val set_priority : t -> Tutils.priority -> unit

Change the process' asynchronous task priority. Useful when switching from blocking read to non-blocking read.

val stop : t -> unit

Asynchronous stop. The process' stdin will be closed some time in the future.

val kill : t -> unit

Asynchronous kill. The process will be killed some time in the future.

val stopped : t -> bool

Whether the process was stopped.

val really_write : ?offset:int -> ?length:int -> bytes -> push -> unit

Write bytes recursively.

val on_stdin : t -> (push -> 'a) -> 'a

Synchronous (blocking) write on the process' stdin. Raises Finished if the process has been stopped/killed.

val on_stdout : t -> (pull -> 'a) -> 'a

Synchronous (blocking) read on the process' stdout. Raises Finished if the process has been stopped/killed.

val on_stderr : t -> (pull -> 'a) -> 'a

Synchronous (blocking) read on the process' stdout. Raises Finished if the process has been stopped/killed.