Module Request

A request is something from which we can produce a file.

type indicator

An indicator is a resource location (URI), when meaningful, it can be declared as temporary if liquidsoap should destroy it after usage (this means deleting a local file).

val conf : Dtools.Conf.ut

Root configuration node.

val indicator : ?metadata:Frame.metadata -> ?temporary:bool -> string -> indicator

Create an indicator.

val pretty_date : Unix.tm -> string

Return a prettified string.

type t

Type of requests, which are devices for obtaining a local file from an URI.

val create : ?resolve_metadata:bool -> ?excluded_metadata_resolvers:string list -> ?metadata:Frame.Metadata.t -> ?persistent:bool -> ?temporary:bool -> cue_in_metadata:string option -> cue_out_metadata:string option -> string -> t

Create a request.

val initial_uri : t -> string

Return the request's initial uri.

val destroy : ?force:bool -> t -> unit

Destroying of a requests causes its file to be deleted if it's a temporary one, for example a downloaded file. If the metadata "persistent" is set to "true", destroying doesn't happen, unless force is set too. Persistent sources are useful for static URIs (see below for the definition of staticity, and in src/sources/one_file.ml for an example of use).

type resolving

type status = [

| `Idle

| `Resolving of resolving

| `Ready

| `Destroyed

| `Failed

]

Status of a request.

val status : t -> status

Current status of a request.

General management

val cleanup : unit -> unit

Called at exit, for cleaning temporary files and destroying all the requests, even persistent ones.

val id : t -> int

Identifier of a request.

val all : unit -> t list

Get the list of all requests.

val from_id : int -> t option

Retrieve a request from its id.

Resolving

Resolving consists in many steps. Every step consist in rewriting the first URI into other URIs. The process ends when the last URI is a local filename. For example, the initial URI can be a database query, which is then turned into a remote locations, which is then tentatively downloaded... At each step protocol.resolve uri timeout is called, and the function is expected to push the new URIs in the request.

type resolver = string -> log:(string -> unit) -> float -> indicator option

Something that resolves an URI.

type protocol = {

resolve : resolver;

static : string -> bool;

}

A protocol, which can resolve associated URIs.

val is_static : string -> bool

A static request r is such that every resolving leads to the same file. Sometimes, it allows removing useless destroy/create/resolve.

type resolve_flag = [

| `Resolved

| `Failed

| `Timeout

]

Resolving can fail because an URI is invalid, or doesn't refer to a valid * audio file, or simply because there was no enough time left.

val conf_metadata_decoder_priorities : Dtools.Conf.ut

Metadata resolvers priorities.

val read_metadata : t -> unit

Read the request's metadata.

val resolve : ?timeout:float -> t -> resolve_flag

resolve ?timeout request tries to resolve the request within timeout seconds. Defaults to settings.request.timeout when timeout is not passed.

val resolved : t -> bool

resolved r if there's an available local filename. It can be true even if the resolving hasn't been run, if the initial URI was already a local filename.

val get_filename : t -> string option

Return a valid local filename if there is one, which means that the request is ready.

Metadatas

val metadata : t -> Frame.metadata

Metadata are resolved from the first indicator to the last, the last one overriding the ones before. The only exception are root metadata, which are metadata internal to liquidsoap such as request id and etc. These cannot be overridden by resolvers.

Logging

Every request has a separate log in which its history can be written.

val log : t -> string

Media operations

These operations are only meaningful for media requests, and might raise exceptions otherwise.

val conf_dresolvers : string list Dtools.Conf.t

Duration resolvers.

val duration : ?resolvers:string list -> metadata:Frame.metadata -> string -> float option

duration ?resolvers ~metadata filename computes the duration of audio data contained in filename. The computation may be expensive. Set resolvers to a list of specific decoders to use for getting duration.

raises Not_found
if no duration computation method is found.

val has_decoder : ctype:Frame.content_type -> t -> bool

true is a decoder exists for the given content-type.

val get_decoder : ctype:Frame.content_type -> t -> Decoder.file_decoder_ops option

Return a decoder if the file has been resolved, guaranteed to have available data to deliver.

val on_air : source:Source.source -> t -> unit

Mark the request as being on_air for the given source.

val done_playing : source:Source.source -> t -> unit

Mark the request as being done playing by the given source.

Plugs

type dresolver = {

dpriority : unit -> int;

file_extensions : unit -> string list;

dresolver : metadata:Frame.metadata -> string -> float;

}

val dresolvers : dresolver Plug.t

Functions for computing duration.

type metadata_resolver = {

priority : unit -> int;

resolver : metadata:Frame.metadata -> extension:string option -> mime:string -> string -> (string * string) list;

}

Type for a metadata resolver. Resolvers are executed in priority order and the first returned metadata take precedence over any other one later returned.

val mresolvers : metadata_resolver Plug.t

Functions for resolving metadata. Metadata filling isn't included in Decoder because we want it to occur immediately after request resolution.

val resolve_metadata : initial_metadata:Frame.metadata -> excluded:string list -> string -> Frame.metadata

Resolve metadata for a local file:

val protocols : protocol Plug.t

Functions for resolving URIs.

module Value : Value.Custom with type content := t