Class Request_dynamic.dynamic

inherit Source.source

method fallible : bool

val mutable remaining : int

method remaining : int

val mutable first_fill : bool

val mutable current : handler option Stdlib.Atomic.t

method current : handler option

method self_sync : Clock.self_sync

val should_skip : bool Stdlib.Atomic.t

method private end_request : unit

How to unload a request.

method private fetch_request : bool

method private generate_from_current_request : int -> Frame.t

method private generate_frame : Frame.t

method seek : int -> int

method seek_source : Source.source

method abort_track : unit

method private is_request_ready : bool

method can_generate_frame : bool

val retrieved : queue_item Queue.t

method private queue_size : int

method queue : queue_item Queue.t

method set_queue : Request.t list -> unit

method add : queue_item -> bool

val state : [> `Sleeping | `Started of float * task ] Stdlib.Atomic.t

method private clear_retrieved : unit

method private notify_new_request : unit

This method should be called whenever the feeding task has a new opportunity to feed the queue, in case it is sleeping.

method private feed_queue : unit -> float

The body of the feeding task

method private synchronous_feed_queue : unit

method fetch : [ `Finished | `Retry ]

method private get_next_file : [ `Empty | `Request of Request.t ]

Provide the unqueued super with resolved requests.