Module Server

Where the server commands are registered, and the server interfaces are * created and controlled.

Configuration parameters

val conf : Dtools.Conf.ut

The configuration of the server.

val conf_socket : bool Dtools.Conf.t

Should a socket server be created?

val conf_socket_path : string Dtools.Conf.t

Path of the socket.

val conf_socket_perms : int Dtools.Conf.t

Permissions of the socket. *

val conf_telnet : bool Dtools.Conf.t

Should a telnet server be created?

val conf_telnet_bind_addr : string Dtools.Conf.t

Address of the telnet server.

val conf_telnet_port : int Dtools.Conf.t

Port of the telnet server.

Commands of the server

type namespace = string list

Namespaces are used to avoid name conflicts.

val to_string : namespace -> string

Get a string representation of a namespace.

type condition = {

wait : (unit -> string) -> unit;

signal : unit -> unit;

broadcast : unit -> unit;

}

Specialized implementation of conditions to use in server commands.

val condition : unit -> condition

condition () instantiates a server command condition.

val write : after:(unit -> string) -> string -> unit

Partial response write without returning.

val read : after:(string -> string) -> Duppy.Io.marker -> unit

Read from the client.

val add : ns:namespace -> ?usage:string -> descr:string -> string -> (string -> string) -> unit

add ~ns ~descr command f adds a new command command in a given namespace ~ns. * When the command is called, the function f is executed with the argument of * the command as parameter. The return value of f is then displayed. ~descr is * the command description.

val remove : ns:namespace -> string -> unit

Remove a command from the server.

val exec : string -> string

exec command executes the command, returns its result. *

raises Not_found
if the command does not exist.

val on_start : (unit -> unit) -> unit

Register a handle to be executed on server start or immediately if the server has already started.

val start : unit -> unit

Start server and telnet if enabled.