Module Source

type sync = [

| `Auto

| `CPU

| `None

]

In `CPU mode, synchronization is governed by the CPU clock. * In `None mode, there is no synchronization control. Latency in * is governed by the time it takes for the sources to produce and * output data. * In `Auto mode, synchronization is governed by the CPU unless at * least one active source is declared self_sync in which case latency * is delegated to this source. A typical example being a source linked * to a sound card, in which case the source latency is governed * by the sound card's clock. Another case is synchronous network * protocol such as input.srt.

type active = < reset : unit ; output : unit >

A source can be: passive, active or output. Active sources and outputs are animated on each clock cycle. Output are kept around until they are manually shutdown. Active and passive sources can be garbage collected if they are not connected to any output. The output method is called on each clock cycle on active sources and outputs. The reset method is called when there is too much latency.

type source_type = [

| `Passive

| `Active of active

| `Output of active

]

exception Unavailable

type streaming_state = [

| `Pending

| `Unavailable

| `Ready of unit -> unit

| `Done of Frame.t

]

Instrumentation.

type metadata = (int * Frame.metadata) list

type watcher = {

wake_up : fallible:bool -> source_type:source_type -> id:string -> ctype:Frame.content_type -> clock_id:string -> unit;

sleep : unit -> unit;

generate_frame : start_time:float -> end_time:float -> length:int -> has_track_mark:bool -> metadata:metadata -> unit;

before_streaming_cycle : unit -> unit;

after_streaming_cycle : unit -> unit;

}

class virtual source : ?stack:Pos.t list -> ?clock:Clock.t -> ?name:string -> unit -> object ... end

The source use is to send data frames through the get method.

class virtual active_source : ?stack:Pos.t list -> ?clock:Clock.t -> ?name:string -> unit -> object ... end

class virtual operator : ?stack:Pos.t list -> ?clock:Clock.t -> ?name:string -> source list -> object ... end

class virtual active_operator : ?stack:Pos.t list -> ?clock:Clock.t -> ?name:string -> source list -> object ... end

type reselect = [

| `Ok

| `Force

| `After_position of int

]

Type governing whether or not the same source should be re-selected.

`Force means that the operator should at least try to see if a new source should be selected.

`Ok means that the operator should re-select the current source whenever possible.

`After_position p means that the operator should re-select the current source only if it can produce data past the given position The generate_from_multiple_sources implements a can_reselect that can be called with the currently selected source to validate this logic.

class virtual generate_from_multiple_sources : merge:(unit -> bool) -> track_sensitive:(unit -> bool) -> unit -> object ... end