Class Source.source

The source use is to send data frames through the get method.

method private mutexify : 'a 'b. ('a -> 'b) -> 'a -> 'b

Naming

method id : string

Identifier of the source.

method set_name : string -> unit

method set_id : ?definitive:bool -> string -> unit

method pos : Pos.Option.t

Position in script

method stack : Pos.t list

method set_stack : Pos.t list -> unit

method stack_unifier : Pos.t list Unifier.t

Source characteristics

method virtual fallible : bool

If false, is_ready should always return true.

method source_type : source_type

The source type. Defaults to `Passive

method active : bool

true if the source needs to be animated on each clock tick.

Init/shutdown

method on_sleep : (unit -> unit) -> unit

Register a callback, to be executed when source shuts down.

method clock : Clock.t

The clock under which the source will run.

method virtual self_sync : Clock.self_sync

Does the source provide its own synchronization? Examples: Alsa, AO, SRT I/O, etc.. This information is used at the clock level to decide whether or not we should synchronize with the CPU clock after producing a frame (for `Auto clocks). Please note that in the case of multiple sources filling the frame with different notion notion of synchronization, there is no consistent notion of time or synchronization. In this case (and with a `Auto clock), we simply decide based on whether there is one self_sync source or not. This logic should dictate how the method is implemented by the various operators.

method source_sync : bool -> Clock.sync_source option

method on_wake_up : (unit -> unit) -> unit

Register a callback when wake_up is called.

method wake_up : unit

Called when the source must be ready. Can be called multiple times

method on_sleep : (unit -> unit) -> unit

Register a callback when sleep is called.

method sleep : unit

Called when the source can release all its resources. Can be called concurrently and multiple times.

method force_sleep : unit

Force the source to sleep. Should be called by the clocks only.

method is_up : bool

Check if a source is up or not.

Streaming

method frame_type : Type.t

method content_type_computation_allowed : unit

This is called when content-type can be computed, i.e. either after frame type has been passed from the typing system during `check_eval` or at `wake_up`

method has_content_type : bool

method content_type : Frame.content_type

What type of content does this source produce.

method private audio_channels : int

This method fails when content is not PCM.

method private video_dimensions : int * int

method buffer : Generator.t

A buffer that can be used by the source.

method private generate_video : ?create: (pos:int -> width:int -> height:int -> unit -> Mm.Video.Canvas.image) -> field:Frame.Fields.field -> int -> Content.Video.data

method last_image : Frame.Fields.field -> Mm.Video.Canvas.image

method private nearest_image : pos:int -> last_image:Mm.Video.Canvas.image -> Content.Video.data -> Mm.Video.Canvas.image

method empty_frame : Frame.t

An empty frame that can be used by the source.

method end_of_track : Frame.t

An empty frame with a track mark.

method virtual remaining : int

Number of main ticks left in the current track. Defaults to -1=unknown.

method elapsed : int

method duration : int

method virtual seek_source : source

Return the source effectively used to seek. Used by the muxer to determine if there is a unique seeking source. Should return self if there isn't such a unique source.

method seek : int -> int

self#seek_ticks x skips x main ticks. returns the number of ticks actually skipped. By default it always returns 0, refusing to seek at all. That method may be called from any thread, concurrently with #get_frame, so they should not interfere.

method last_metadata : Frame.metadata option

The source's last metadata.

method reset_last_metadata_on_track : bool

method set_reset_last_metadata_on_track : bool -> unit

method register_command : ?usage:string -> descr:string -> string -> (string -> string) -> unit

Register a server command. The command is registered when the source wakes up under its own id as namespace and deregistered when it goes down.

method on_metadata : (Frame.metadata -> unit) -> unit

Register a callback to be called on new metadata

method on_track : (Frame.metadata -> unit) -> unit

Register a callback to be called on new track. Callback is called with the most recent metadata before a given track mark.

method private execute_on_track : Frame.t -> unit

These two are used by generate_from_multiple_sources and should not be used otherwise.

method private set_last_metadata : Frame.t -> unit

method private virtual can_generate_frame : bool

Sources must implement this method. It should return true when the source can produce data during the current streaming cycle.

method on_before_streaming_cycle : (unit -> unit) -> unit

method on_after_streaming_cycle : (unit -> unit) -> unit

method private virtual generate_frame : Frame.t

Sources must implement this method. It should return the data produced during the current streaming cycle. Sources are responsible for producing as much data as possible, up-to the frame size setting.

method is_ready : bool

This method is based on can_generate_frame and has the same value through the whole streaming cycle.

method get_frame : Frame.t

If the source is ready, this method computes the frame generated by the source during the current streaming cycle. Returned value is cached and should be the same throughout the whole streaming cycle.

method get_partial_frame : (Frame.t -> Frame.t) -> Frame.t

This method passes the frame returned by #get_frame to the given callback. The callback should return the portion of the frame (of the form: start, end)) that was effectively used. This method is used when a consumer of the source's data only uses an initial chunk of the frame. In this case, the remaining data is cached whenever possible and returned during the next streaming cycle. Final returned value is the same as the partial chunk returned from the callback for easy method call chaining. Calling this method is equivalent to doing: {[ let frame = Frame.slice source#peek_frame len in source#consumed (Frame.position frame); frame ]}

method peek_frame : Frame.t

Check a frame without consuming any of its data.

method consumed : int -> unit

Manually mark amount of consumed data from the source.

method get_mutable_content : Frame.field -> Content.data

This method requests a specific field of the frame that can be mutated. It is used by a consumer of the source that will modify the source's data (e.g. amplify). The source will do its best to minimize data copy according to the streaming context. Typically, if there is only one consumer of the source's data, it should be safe to pass its data without copying it.

method get_mutable_frame : Frame.field -> Frame.t

This method is the same as #get_mutable_content but returns a full frame with the requested mutable field included.

method private split_frame : Frame.t -> Frame.t * Frame.t option

By convention, frames produced during the streaming cycle can only have at most one track mark. In case of multiple track marks (which most likely indicate a programming problem), all subsequent track marks past the first one are dropped.

This function returns a pair: (initial_frame, new_track option) of an initial frame and, if a track mark is present in the frame, the optional portion of the frame contained after this track mark.

This method can be used to implement operations that should be aware of new tracks.

method set_frame_data : Frame.field -> (?offset:int -> ?length:int -> 'a -> Content.data) -> 'a -> Frame.t

This method is a convenience function to set some data. It returns the frame produced by the source during the current streaming cycle with the given field data replaced by the data passed to the function with length set as the frame's length.

Various information related to the current frame.

method frame_position : int

method frame_audio_position : int

method frame_has_track_mark : bool

method frame_track_mark : int option

method frame_metadata : (int * Frame.Metadata.t) list

method virtual abort_track : unit

Tells the source to end its current track.

Utilities

method log : Log.t

method add_watcher : watcher -> unit