Module Strings.Mutable

Mutable and thread-safe variant.

type t

val empty : unit -> t

The empty buffer.

val of_string : string -> t

Create a buffer from a string.

val of_strings : buffer -> t

val of_bytes : bytes -> t

Create a buffer from the given bytes which will be copied.

val unsafe_of_bytes : bytes -> t

Create a buffer from the given bytes which will not be copied (be careful).

val to_string : t -> string

Render a buffer into a string/bytes. This operation can be costly (in terms of memory copies), avoid it.

val to_bytes : t -> bytes

val to_strings : t -> buffer

val substring : t -> int -> int -> string

val of_list : string list -> t

Concatenation of strings.

val add : t -> string -> unit

Add a string at the end of a buffer.

val add_bytes : t -> bytes -> unit

Add bytes at the end of a buffer, bytes will be copied.

val unsafe_add_bytes : t -> bytes -> unit

Add bytes at the end of a buffer, bytes will not be copied.

val add_substring : t -> string -> int -> int -> unit

val add_subbytes : t -> bytes -> int -> int -> unit

Add subbytes at the end of a buffer.

val unsafe_add_subbytes : t -> bytes -> int -> int -> unit

Add subbytes at the end of a buffer with copying them.

val dda : string -> t -> unit

Add a string at the beginning of a buffer.

val iter : (string -> int -> int -> unit) -> t -> unit

Iterate a function on all the strings (with given offset and length) contained in the buffer.

val iter_view : (StringView.t -> unit) -> t -> unit

val fold : ('a -> string -> int -> int -> 'a) -> 'a -> t -> 'a

Fold a function over all the strings (with given offset and length) contained in the buffer.

val fold_view : ('a -> StringView.t -> 'a) -> 'a -> t -> 'a

val map : (string -> int -> int -> string * int * int) -> t -> t

Map a function over all the strings (with given offset and length) contained in the buffer.

val map_view : (StringView.t -> StringView.t) -> t -> t

val drop : t -> int -> unit

Drop the first given bytes.

val keep : t -> int -> unit

Keep the last given bytes.

val sub : t -> int -> int -> t

Sub-buffer of a buffer.

val blit : t -> int -> bytes -> int -> int -> unit

Copy a substring.

val is_empty : t -> bool

Whether the buffer is the empty string.

val length : t -> int

Length of the buffer.

val append : t -> t -> unit

Append two buffers.

val append_strings : t -> buffer -> unit

Append strings to the buffer.

val flush : t -> buffer

Empty the buffer and return its content.