Module Switch

Custom operator which selects one of its children sources * either at the beginning of a track or at every frame, * depending on a parametrizable predicate. * A few specializations of it are defined below.

type transition = Lang.value

type child = {

source : Source.source;

transition : transition;

}

type track_mode =

| Sensitive

| Insensitive

The switch can either happen at any time in the stream (insensitive) * or only at track limits (sensitive).

type selection = {

predicate : Lang.value;

child : child;

effective_source : Source.source;

}

val satisfied : Lang.value -> bool

val trivially_true : Value.t -> bool

val pick_selection : ('a * 'b * 'c) -> 'a * 'c

val find : ?strict:bool -> ('a -> bool) -> 'a list -> 'a

Like List.find but evaluates f on every element when strict is true.

class switch : all_predicates:bool -> override_meta:string -> transition_length:int -> replay_meta: bool -> track_sensitive:(unit -> bool) -> (Value.t * bool * child) list -> object ... end

Common tools for Lang bindings of switch operators

val default_transition : Lang.value