Module Time_warp.AdaptativeBuffer

module RB : sig ... end

Ringbuffers where number of channels is fixed on first write.

type control = {

lock : Mutex.t;

rb : RB.t;

mutable rb_length : float;

mutable mg : Generator.t;

mutable buffering : bool;

mutable abort : bool;

}

The kind of value shared by a producer and a consumer.

val proceed : control -> (unit -> 'a) -> 'a

class producer : pre_buffer:float -> averaging:float -> limit:float -> resample:bool -> control -> object ... end

The source which produces data by reading the buffer.

class consumer : autostart:bool -> infallible:bool -> on_start:(unit -> unit) -> on_stop:(unit -> unit) -> pre_buffer:float -> reset:bool -> Lang.value -> control -> object ... end

val create : autostart:bool -> infallible:bool -> on_start:(unit -> unit) -> on_stop:(unit -> unit) -> pre_buffer:float -> max_buffer:float -> averaging:float -> limit:float -> reset:bool -> resample:bool -> Lang.value -> producer