Module Time_warp.Buffer

Create a buffer between two clocks. * * This creates an active operator in the inner clock (the action consists * in filling the buffer) but it does not create or force in any * way the clock that's going to animate it. * * We actually create two sources, to avoid the mess of having a same * source belonging to one clock but being animated by another. * This makes it possible to have the inner clock stop and shutdown * the source that feeds the buffer, without disturbing the other clock * in which the buffer-consumer will still behave OK (except obviously * that the buffer will empty).

type control = {

lock : Mutex.t;

generator : Generator.t Stdlib.Lazy.t;

mutable buffering : bool;

mutable abort : bool;

}

val proceed : control -> (unit -> 'a) -> 'a

class producer : id:string -> control -> object ... end

The source which produces data by reading the buffer.

class consumer : id:string -> autostart:bool -> infallible:bool -> on_start:(unit -> unit) -> on_stop:(unit -> unit) -> pre_buffer:float -> max_buffer:float -> Lang.value -> control -> object ... end

val create : id:string -> autostart:bool -> infallible:bool -> on_start:(unit -> unit) -> on_stop:(unit -> unit) -> pre_buffer:float -> max_buffer:float -> Lang.value -> producer