Module Udp_io.Tutils

include module type of struct include Tutils end

val error_handler : bt:string -> exn -> bool

val error_handlers : (bt:string -> exn -> bool) Stdlib.Stack.t

val create : ('a -> unit) -> 'a -> string -> Thread.t

Thread wrapper

Give names to threads, and forbid them to raise an exception; if that happens, the thread dies anyway but it is logged and main will notice it. The main process is expected to run main after having launched the needed threads: that function will sleep until a thread raises an exception.

val main : unit -> unit

val start : unit -> unit

val running : unit -> bool

val finished : unit -> bool

val shutdown : int -> unit

val cleanup : unit -> unit

val exit_code : unit -> int

val exit : unit -> unit

exception Exit

Special exception allowed for "clean" termination of Tutils threads. * All other exceptions are reported as bugs.

val join_all : unit -> unit

Wait for the threads to terminate, * never return if some thread keeps running.

Multi-tasking scheduler

type priority = [

| `Blocking
(*
For example a last.fm submission.
*)

| `Maybe_blocking
(*
Request resolutions vary a lot.
*)

| `Non_blocking
(*
Non-blocking tasks like the server.
*)

]

Priorities for the different scheduler usages.

val scheduler : priority Duppy.scheduler

task scheduler

Misc

val wait : Condition.t -> Mutex.t -> (unit -> bool) -> unit

Waits for f() to become true on condition c. * The mutex m protecting data accessed by f is in the same state before * and after the call.

exception Timeout of float

type event = [

| `Read of Unix.file_descr

| `Write of Unix.file_descr

| `Both of Unix.file_descr

]

val wait_for : ?log:(string -> unit) -> event -> float -> unit

val seems_locked : Mutex.t -> bool

Tests whether a mutex is locked, without blocking. * We cannot check on Win32, where true is always returned: * it always "seems" OK, we don't raise false alarms. * This is meant to be used for assertions.

val write_all : ?timeout:float -> Unix.file_descr -> bytes -> unit

val stoppable_thread : (((unit -> bool) * (unit -> unit)) -> unit) -> string -> (unit -> unit) * (unit -> unit)

Preemptive stoppable thread. * * The thread function receives a should_stop,has_stop pair on startup. * It should regularly poll the should_stop and stop when asked to. * Before stopping it should call has_stopped. * * The function returns a kill,wait pair. The first function should be * called to request that the thread stops, and the second to wait * that it has effectively stopped.