Module Utils

include module type of struct include Liquidsoap_lang.Utils end

val resolve_path : ?current_dir:string -> string -> string

val readable : string -> bool

val check_readable : ?current_dir:string -> pos:Liquidsoap_lang.Pos.t list -> string -> string

val string_of_float : float -> string

val select : Unix.file_descr list -> Unix.file_descr list -> Unix.file_descr list -> float -> Unix.file_descr list * Unix.file_descr list * Unix.file_descr list

val log_exception : log:Log.t -> bt:string -> string -> unit

module Thread : sig ... end

val force_locale : string -> unit

val pagesize : int

val prefix : 'a list -> 'a list -> bool

val remove_one : ('a -> bool) -> 'a list -> 'a list

Remove the first element satisfying a predicate, raising Not_found * if none is found.

val read_all : string -> string

val copy : ?mode:Stdlib.open_flag list -> ?perms:int -> string -> string -> unit

val buffer_drop : Stdlib.Buffer.t -> int -> unit

val unix_translator : exn -> string option

val interpolate : (string -> string) -> string -> string

val which : path:string list -> string -> string

which s is equivalent to /usr/bin/which s, raises Not_found on error

val which_opt : path:string list -> string -> string option

val timezone : unit -> int

Get current timezone.

val timezone_by_name : unit -> string * string

type tm = {

tm_sec : int;

tm_min : int;

tm_hour : int;

tm_mday : int;

tm_mon : int;

tm_year : int;

tm_isdst : bool option;

}

val mktime : tm -> float

val string_of_timezone : int -> string

val strftime : ?time:float -> string -> string

Very partial strftime clone

val is_dir : string -> bool

Check if a directory exists.

val dir_exists : string -> bool

val mkdir : perm:Unix.file_perm -> string -> unit

Create a directory, and its parents if needed. * Raise Unix_error on error.

val get_tempdir : unit -> string

val get_ext : string -> string

Get a file/uri extension.

val get_ext_opt : string -> string option

val name_of_sockaddr : ?rev_dns:bool -> ?show_port:bool -> Unix.sockaddr -> string

val uptime : unit -> float

Uptime since the application was started.

val normalize_parameter_string : string -> string

Generate a string which can be used as a parameter name.

val reopen_out : Stdlib.out_channel -> string -> unit

A function to reopen a file descriptor * Thanks to Xavier Leroy! * Ref: http://caml.inria.fr/pub/ml-archives/caml-list/2000/01/ * a7e3bbdfaab33603320d75dbdcd40c37.en.html

val reopen_in : Stdlib.in_channel -> string -> unit

The same for inchan

val float_of_extended_float : string -> float

val reachable_size : 'a -> int

Size of an OCaml value (including referred elements) in bytes.

val string_of_size : int -> string

String representation of a size (in bytes).

val var_script : string Stdlib.ref

val substs : (string * (unit -> string)) list Stdlib.ref

val add_subst : string -> string -> unit

val subst_vars : string -> string

val concat_with_last : last:string -> string -> string list -> string

val abs_float : float -> float

val frame_id_of_string : string -> Metadata.ID3v2.frame_id option

val id3v2_of_metadata : version:int -> (string * string) list -> string

val is_docker : bool Stdlib.Lazy.t

val optional_apply : ('a -> unit) -> 'a option -> unit