Module Video_converter

val video_converter_conf : Dtools.Conf.ut

Plugin to add video-related configuration keys.

type converter = Mm.Image.Generic.t -> Mm.Image.Generic.t -> unit

f src dst performs the conversion from frame src to frame dst. Raises Not_found if no conversion routine was found.

type converter_plug = Mm.Image.Generic.Pixel.format list * Mm.Image.Generic.Pixel.format list * (unit -> converter)

A converter plugin is a name, a list of input formats, a list of output formats, a function to create a converter.

val video_converters : converter_plug Plug.t

Plugin to register new converters.

val find_converter : Mm.Image.Generic.Pixel.format -> Mm.Image.Generic.Pixel.format -> Mm.Image.Generic.t -> Mm.Image.Generic.t -> unit

find_converter source destination tries to find a converter from source format to destination format. Proportional scale is implicitly set via global configuration key for now. Returns a conversion function: frame -> frame -> unit.

val scaler : unit -> Mm.Video.Image.t -> Mm.Video.Image.t -> unit

Generate a function to scale images.