Module Wav_aiff

Encode/decode IFF files, that is AIFF and WAV.

type 'a t

type format = [

| `Aiff

| `Wav

]

val format_of_handler : 'a t -> format

exception Not_a_iff_file of string

type 'a read_ops = {

really_input : 'a -> Stdlib.Bytes.t -> int -> int -> unit;

input_byte : 'a -> int;

input : 'a -> Stdlib.Bytes.t -> int -> int -> int;

seek : 'a -> int -> unit;

close : 'a -> unit;

}

val in_chan_ops : Stdlib.in_channel read_ops

type callback = Stdlib.Bytes.t -> int -> int -> int

val callback_ops : callback read_ops

val fopen : string -> Stdlib.in_channel t

Open the named file for reading, and return a new wav descriptor. Raise Sys_error if the file could not be opened and Not_a_iff_file if it hasn't the right format.

val read_header : 'a read_ops -> 'a -> 'a t

Generic opener.

val in_chan_read_header : Stdlib.in_channel -> Stdlib.in_channel t

Read data from an input channel.

val info : 'a t -> string

info w returns a string containing some information on wav w

val channels : 'a t -> int

Parameters of the output PCM format.

val sample_rate : 'a t -> int

val sample_size : 'a t -> int

val data_length : 'a t -> int

val close : 'a t -> unit

close w close the wav descriptor w

val wav_header : ?len:int -> channels:int -> sample_rate:int -> sample_size:int -> unit -> string

Returns the WAV header that declares the given format. * The lengths of file and data are set to their maximum possible value.

val duration : 'a t -> float

Returns the duration of the data. Warning: value may not be accurate for streams.