Module Builtins_ffmpeg_filters

module Queue = Liquidsoap_lang.Queues.Queue

FFmpeg filter graphs initialization is pretty tricky. Things to consider: * - FFmpeg filters are using a push paradigm, pushing from the sources down * to the outputs * - Liquidsoap uses a pull paradigm, pulling data from the outputs. * - FFmpeg filters inputs need to know the exact content of the data sent to * them before being initialized, which is only known in the worst case * when receiving the first frame. * * Therefore, the intended implementation is to: * -> Consider ffmpeg filter graphs as a single operator with N inputs * and M outputs (audio/video) with inputs being any source converted to * a ffmpeg graph input, even if not used, for simplification. * -> The outputs are placed in their clock which controls a child clock * containing the inputs. * -> The graph initialization is suspended until its initialization conditions * are met. * -> When all the inputs have been initialized and are ready (call to `#is_ready`), * the graph is considered ready and its output can start requesting * content. This is captured by the call to `is_ready` below. * -> When requesting content, the outputs can tick the input clock as many time * as needed to generate output data. We expect more or less real time with * perhaps an accordion pattern between input and output so we do not * look at latency control like we do for crossfades. * -> When receiving its first frame, the liquidsoap input will then initialize the * corresponding ffmpeg graph input with full format info. * -> When all inputs have been initialized, which is know by checking if all of * the graph's lazy values for input pads have been forced (executed), * the whole graph is initialized, outputs connected and data can start to flow! * This is captured by the call to `initialized` below. * * This is contingent to the inputs being checked for `#is_ready` and the * outputs only pulling when _all_ inputs are available, to avoid running * into endless pulling loops.

val ffmpeg_filter_audio : Lang.module_name

val ffmpeg_filter_video : Lang.module_name

val log : Log.t

type 'a input = 'a Avfilter.input

type 'a output = 'a Avfilter.output

type 'a setter = 'a -> unit

type 'a entries = (string, 'a setter) Stdlib.Hashtbl.t

type inputs = ([ `Audio ] input entries, [ `Video ] input entries) Avfilter.av

type outputs = ([ `Audio ] output entries, [ `Video ] output entries) Avfilter.av

type graph = {

mutable config : Avfilter.config option;

mutable failed : bool;

input_inits : (unit -> bool) Queue.t;

graph_inputs : Source.source Queue.t;

graph_outputs : Source.source Queue.t;

init : unit Stdlib.Lazy.t Queue.t;

entries : (inputs, outputs) Avfilter.io;

}

val init_graph : graph -> unit

val initialized : graph -> bool

val is_ready : graph -> bool

val pull : graph -> unit

val self_sync : graph -> < id : string ; stack : Pos.t list.. > -> [> `Dynamic | `Static ] * Clock_base.sync_source option

module Graph : sig ... end

module Audio : sig ... end

module Video : sig ... end

val uniq_name : string -> string

exception No_value_for_option

val mk_options : Avutil.Options.t -> (string * Lang.t * Lang.value option * string option) list * ((string * Lang.value) list -> [> `Pair of string * [> `Int64 of int64 | `String of string ] ] as 'a list -> 'a list)

val get_config : Liquidsoap_lang.Value.t -> Avfilter.config

val apply_filter : args_parser: ((string * Liquidsoap_lang.Value.t) list -> 'a list -> Avfilter.args list) -> filter:[ `Unattached ] Avfilter.filter -> sources_t:('b * string * 'c) list -> (string * Liquidsoap_lang.Value.t) list -> Lang.value

val register_filters : unit -> unit

val abuffer_args : Avutil.audio Avutil.frame -> [> `Pair of string * [> `Int of int | `Int64 of int64 | `Rational of Avutil.rational ] ] list

val buffer_args : Avutil.video Avutil.frame -> [> `Pair of string * [> `Int of int | `Rational of Avutil.rational ] ] list

val unify_clocks : clock:Clock.t -> < clock : Clock.t ; pos : Liquidsoap_lang.Pos.Option.t.. > Queue.t -> unit