Module Ladspa_op

module Cache = Liquidsoap_lang.Cache

module String : sig ... end

val ladspa : Lang.module_name

type t =

| Float

| Int

| Bool

val log : Log.t

val ladspa_enabled : bool

val ladspa_dirs : string list

val port_t : Ladspa.Descriptor.t -> int -> t

class virtual base : Source.source -> object ... end

class virtual base_nosource : object ... end

val instantiate : Ladspa.Descriptor.t -> int -> Ladspa.Descriptor.instance

class ladspa_mono : Source.source -> string -> int -> int -> int -> (int * (unit -> float)) list -> object ... end

class ladspa : Source.source -> string -> int -> int array -> int array -> (int * (unit -> float)) list -> object ... end

class ladspa_nosource : string -> int -> int array -> (int * (unit -> float)) list -> object ... end

type port = {

port_index : int;

port_name : string;

port_type : t;

port_default : float option;

port_min : float option;

port_max : float option;

}

val default_samplerate : int

When creating operator for LADSPA plugins, we don't know yet at which samplerate Liquidsoap will operate. But the default values and bounds for LADSPA parameters might depend on the samplerate. Lacking a better solution, we use the following default samplerate, potentially creating a mismatch between the doc and the actual behavior.

val get_control_ports : Ladspa.Descriptor.t -> port list

type plugin = {

plugin_file : string;

plugin_descriptor : int;

plugin_inputs : int array;

plugin_outputs : int array;

plugin_controls : port list;

plugin_name : string;

plugin_label : string;

plugin_maker : string;

}

val get_audio_ports : Ladspa.Descriptor.t -> int array * int array

Get input and output ports.

val load_descriptor : string -> int -> Ladspa.Descriptor.t -> plugin

val params_of_controls : port list -> (string * Lang.t * Lang.value option * string option) list * ((string * Lang.value) list -> (int * (unit -> float)) list)

val register_descr : plugin -> unit

val register_plugin : plugin Cache.Table.t -> string -> unit

val register_plugins : unit -> unit