Module Xmlplaylist

Generic xml playlist parsing module

This module implements basic playlist parsing for various xml formats.

Currently supported formats are: podcast (rss), xspf, smil asx

Types and exceptions

type error =

| XmlError of string

| Empty

| UnknownType

| Internal

exception Error of error

type format =

| Podcast

| Xspf

| Smil

| Asx

Functions

val string_of_error : error -> string

Get meaning of Error e

val tracks : ?format:format -> string -> ((string * string) list * string) list

tracks data performs whole process and outputs a list of metadatas,uri from given xml data string

All metadatas are what is provided by the playlist The only variable name that is changed is the author because each formats has its own field for that. The module will use "artist".

Order of tracks is preserved.

format is an optional argument, used to force format detection. Otherwise. detect_format is used.

val detect_format : string -> format

Try to detect the format automatically.