Module Lang_source

module Lang = Liquidsoap_lang.Lang

module Flags = Liquidsoap_lang.Flags

module ClockValue : sig ... end

val log : Log.t

val metadata_t : Lang.t

val to_metadata_list : Lang.value -> (string * string) list

val to_metadata : Lang.value -> Frame.Metadata.t

val metadata_list : (string * string) list -> Lang.value

val metadata : Frame.Metadata.t -> Lang.value

val metadata_track_t : Type.t

val track_marks_t : Type.t

module Source_val : sig ... end

val make_t : ?pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> Type.descr -> Type.t

val source_methods_t : Lang.t -> Lang.t

val source_t : ?pos:Liquidsoap_lang.Term_base.parsed_pos -> ?methods:bool -> Type.t -> Lang.t

val of_source_t : Type.t -> Type.t

val source_tracks_t : Type.t -> Type.t

val source_methods : base:Lang.value -> < abort_track : unit ; active : bool ; add_watcher : Source.watcher -> unit ; buffer : Generator.t ; clock : ClockValue.content ; consumed : int -> unit ; content_type : (Frame.Fields.field, Content.format) Liquidsoap_lang__Methods.t ; content_type_computation_allowed : unit ; duration : int ; elapsed : int ; empty_frame : Frame.t ; end_of_track : Frame.t ; fallible : bool ; force_sleep : unit ; frame_audio_position : int ; frame_has_track_mark : bool ; frame_metadata : (int * Frame.Metadata.t) list ; frame_position : int ; frame_track_mark : int option ; frame_type : Type.t ; get_frame : Frame.t ; get_mutable_content : Frame.field -> Content.data ; get_mutable_frame : Frame.field -> Frame.t ; get_partial_frame : (Frame.t -> Frame.t) -> Frame.t ; has_content_type : bool ; id : string ; is_ready : bool ; is_up : bool ; last_image : Frame.Fields.field -> Mm.Video.Canvas.image ; last_metadata : Frame.Metadata.t option ; log : < active : int -> bool ; critical : 'a. ('a, unit, string, unit) Stdlib.format4 -> 'a ; debug : 'a. ('a, unit, string, unit) Stdlib.format4 -> 'a ; f : 'a. int -> ('a, unit, string, unit) Stdlib.format4 -> 'a ; important : 'a. ('a, unit, string, unit) Stdlib.format4 -> 'a ; info : 'a. ('a, unit, string, unit) Stdlib.format4 -> 'a ; level : int option ; set_level : int -> unit ; severe : 'a. ('a, unit, string, unit) Stdlib.format4 -> 'a > ; on_after_streaming_cycle : (unit -> unit) -> unit ; on_before_streaming_cycle : (unit -> unit) -> unit ; on_metadata : (Frame.Metadata.t -> unit) -> unit ; on_sleep : (unit -> unit) -> unit ; on_track : (Frame.Metadata.t -> unit) -> unit ; on_wake_up : (unit -> unit) -> unit ; peek_frame : Frame.t ; pos : Pos.Option.t ; register_command : ?usage:string -> descr:string -> string -> (string -> string) -> unit ; remaining : int ; reset_last_metadata_on_track : bool ; seek : int -> int ; seek_source : Source.source ; self_sync : [ `Dynamic | `Static ] * Clock.sync_source option ; set_frame_data : 'a. Frame.field -> (?offset:int -> ?length:int -> 'a -> Content.data) -> 'a -> Frame.t ; set_id : ?definitive:bool -> string -> unit ; set_name : string -> unit ; set_reset_last_metadata_on_track : bool -> unit ; set_stack : Pos.t list -> unit ; sleep : unit ; source_sync : bool -> Clock.sync_source option ; source_type : Source.source_type ; stack : Pos.t list ; stack_unifier : Pos.t list Unifier.t ; wake_up : unit > -> Lang.value

val source : Source_val.content -> Lang.value

val track : Track.content -> Value.t

val to_source : Liquidsoap_lang__Value.t -> Source_val.content

val to_source_list : Lang.value -> Source_val.content list

val to_track : Value.t -> Track.content

type 'a operator_method = string * Lang.scheme * string * ('a -> Lang.value)

A method: name, type scheme, documentation and implementation (which takes the currently defined source as argument).

val has_value_flag : Value.t -> Liquidsoap_lang.Flags.flag -> bool

val add_value_flag : Value.t -> Liquidsoap_lang.Flags.flag -> unit

val check_content : Value.t -> Type.t -> unit

Ensure that the frame contents of all the sources occurring in the value agree with t.

An operator is a builtin function that builds a source. * It is registered using the wrapper add_operator. * Creating the associated function type (and function) requires some work: * - Specify which content_kind the source will carry: * a given fixed number of channels, any fixed, a variable number? * - The content_kind can also be linked to a type variable, * e.g. the parameter of a format type. * From this high-level description a type is created. Often it will * carry a type constraint. * Once the type has been inferred, the function might be executed, * and at this point the type might still not be known completely * so we have to force its value within the acceptable range.

val _meth : Lang.value -> (string * Lang.value) list -> Lang.value

val check_arguments : env:(string * Liquidsoap_lang.Value.t) list -> return_t:Liquidsoap_lang.Type_base.t -> (string * Liquidsoap_lang.Type_base.t * 'a * 'b) list -> Liquidsoap_lang.Type_base.t * (string * Liquidsoap_lang.Value.t) list

val add_operator : category:Doc.Value.source -> descr:string -> ?flags:Liquidsoap_lang.Doc.Value.flag list -> ?meth: < abort_track : unit ; active : bool ; add_watcher : Source.watcher -> unit ; buffer : Generator.t ; clock : Clock.t ; consumed : int -> unit ; content_type : Frame.content_type ; content_type_computation_allowed : unit ; duration : int ; elapsed : int ; empty_frame : Frame.t ; end_of_track : Frame.t ; fallible : bool ; force_sleep : unit ; frame_audio_position : int ; frame_has_track_mark : bool ; frame_metadata : (int * Frame.Metadata.t) list ; frame_position : int ; frame_track_mark : int option ; frame_type : Type.t ; get_frame : Frame.t ; get_mutable_content : Frame.field -> Content.data ; get_mutable_frame : Frame.field -> Frame.t ; get_partial_frame : (Frame.t -> Frame.t) -> Frame.t ; has_content_type : bool ; id : string ; is_ready : bool ; is_up : bool ; last_image : Frame.Fields.field -> Mm.Video.Canvas.image ; last_metadata : Frame.metadata option ; log : Log.t ; on_after_streaming_cycle : (unit -> unit) -> unit ; on_before_streaming_cycle : (unit -> unit) -> unit ; on_metadata : (Frame.metadata -> unit) -> unit ; on_sleep : (unit -> unit) -> unit ; on_track : (Frame.metadata -> unit) -> unit ; on_wake_up : (unit -> unit) -> unit ; peek_frame : Frame.t ; pos : Pos.Option.t ; register_command : ?usage:string -> descr:string -> string -> (string -> string) -> unit ; remaining : int ; reset_last_metadata_on_track : bool ; seek : int -> int ; seek_source : Source.source ; self_sync : Clock.self_sync ; set_frame_data : 'a. Frame.field -> (?offset:int -> ?length:int -> 'a -> Content.data) -> 'a -> Frame.t ; set_id : ?definitive:bool -> string -> unit ; set_name : string -> unit ; set_reset_last_metadata_on_track : bool -> unit ; set_stack : Pos.t list -> unit ; sleep : unit ; source_sync : bool -> Clock.sync_source option ; source_type : Source.source_type ; stack : Pos.t list ; stack_unifier : Pos.t list Unifier.t ; wake_up : unit.. > as 'b operator_method list -> ?base:Lang.module_name -> string -> (string * Lang.t * Lang.value option * string option) list -> return_t:Liquidsoap_lang.Type_base.t -> ((string * Liquidsoap_lang.Value.t) list -> 'c) -> Lang.module_name

val add_track_operator : category:Doc.Value.source -> descr:string -> ?flags:Liquidsoap_lang.Doc.Value.flag list -> ?meth: < abort_track : unit ; active : bool ; add_watcher : Source.watcher -> unit ; buffer : Generator.t ; clock : Clock.t ; consumed : int -> unit ; content_type : Frame.content_type ; content_type_computation_allowed : unit ; duration : int ; elapsed : int ; empty_frame : Frame.t ; end_of_track : Frame.t ; fallible : bool ; force_sleep : unit ; frame_audio_position : int ; frame_has_track_mark : bool ; frame_metadata : (int * Frame.Metadata.t) list ; frame_position : int ; frame_track_mark : int option ; frame_type : Type.t ; get_frame : Frame.t ; get_mutable_content : Frame.field -> Content.data ; get_mutable_frame : Frame.field -> Frame.t ; get_partial_frame : (Frame.t -> Frame.t) -> Frame.t ; has_content_type : bool ; id : string ; is_ready : bool ; is_up : bool ; last_image : Frame.Fields.field -> Mm.Video.Canvas.image ; last_metadata : Frame.metadata option ; log : Log.t ; on_after_streaming_cycle : (unit -> unit) -> unit ; on_before_streaming_cycle : (unit -> unit) -> unit ; on_metadata : (Frame.metadata -> unit) -> unit ; on_sleep : (unit -> unit) -> unit ; on_track : (Frame.metadata -> unit) -> unit ; on_wake_up : (unit -> unit) -> unit ; peek_frame : Frame.t ; pos : Pos.Option.t ; register_command : ?usage:string -> descr:string -> string -> (string -> string) -> unit ; remaining : int ; reset_last_metadata_on_track : bool ; seek : int -> int ; seek_source : Source.source ; self_sync : Clock.self_sync ; set_frame_data : 'a. Frame.field -> (?offset:int -> ?length:int -> 'a -> Content.data) -> 'a -> Frame.t ; set_id : ?definitive:bool -> string -> unit ; set_name : string -> unit ; set_reset_last_metadata_on_track : bool -> unit ; set_stack : Pos.t list -> unit ; sleep : unit ; source_sync : bool -> Clock.sync_source option ; source_type : Source.source_type ; stack : Pos.t list ; stack_unifier : Pos.t list Unifier.t ; wake_up : unit.. > as 'b operator_method list -> ?base:Lang.module_name -> string -> (string * Lang.t * Lang.value option * string option) list -> return_t:Liquidsoap_lang.Type_base.t -> ((string * Liquidsoap_lang.Value.t) list -> Frame.Fields.field * 'c) -> Lang.module_name

val iter_sources : ?on_imprecise:(unit -> unit) -> (Source_val.content -> unit) -> Value.t -> unit